requirements.dev.txt
Django==1.6.2
argparse==1.1
coverage==3.7.1
dj-database-url==0.2.2
dj-static==0.0.5
django-datetime-widget==0.6
django-haystack==2.1.0
gunicorn==18.0
pystache==0.5.3
pytz==2014.2
requests==2.2.1
selenium==2.40.0
static==1.0.2
wsgiref==0.1.2
